requirements.txt
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
binaryornot==0.4.4
certifi==2026.2.25
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.7
click==8.3.2
cookiecutter==2.7.1
flet-cli==0.84.0
flet-desktop==0.84.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
oauthlib==3.3.1
packaging==26.1
psycopg2-binary==2.9.11
pycparser==3.0
pyfiglet==1.0.4
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
PyYAML==6.0.3
qrcode==8.2
repath==0.9.0
requests==2.33.1
rich==15.0.0
six==1.17.0
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
watchdog==6.0.0
